댓글 권한 완료

+++ EatPlatform/src/main/java/com/eatplatform/web/controller/ReviewRESTController.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -98,11 +99,19 @@ public class ReviewRESTController {
 	// 리뷰 삭제
 	@DeleteMapping("/{reviewId}")
 	public ResponseEntity<Integer> deleteReview(
-			@PathVariable("reviewId") int reviewId) {
-		log.info("deleteReview()");
+			@PathVariable("reviewId") int reviewId,
+			@AuthenticationPrincipal UserDetails userDetails) {
 		
+		String loggedInUserId = userDetails.getUsername();
+		ReviewVO reviewVO = reviewService.getReviewById(reviewId);
+		String reviewUserId = reviewVO.getUserId();
+		
+		if(loggedInUserId.equals(reviewUserId)) {
+		
+		log.info("deleteReview()");
 		int result = reviewService.deleteReview(reviewId);
-		return new ResponseEntity<Integer>(result, HttpStatus.OK);
+		}
+		return new ResponseEntity<Integer>(HttpStatus.OK);
 	}
 	
 }
